+++ components/AboutIndex/content.ts
