+++ components/Footer/content.ts
