+++ components/Footer/index.tsx
@@ -1,10 +1,9 @@
 import { useContext } from 'react';
 import { Divider, Text, Anchor } from '@mantine/core';
 import { LanguageContext } from '../../context';
-import { languageOptions } from '../../models';
 import { labels } from './content';
 
-export const Footer = () => {
+export default function Footer() {
   const { language } = useContext(LanguageContext);
   return (
     <div style={{ marginTop: 'auto', padding: '1rem 0rem' }}>
@@ -13,4 +12,4 @@ export const Footer = () => {
       <Anchor href="#">us</Anchor>
     </div>
   );
-};
+}
