+++ components/Table/content.ts
