+++ components/index.tsx
@@ -1,10 +1,11 @@
-export { Nav } from './nav';
-export { Footer } from './footer';
-export { UsersLookup } from './usersLookup/usersLookup';
-export { Layout } from './layout';
-export { SignupForm } from './signupForm/signupForm';
-export { Table } from './table/table';
-export { Modal } from './modal/modal';
-export { AboutIndex } from './aboutIndex/aboutIndex';
-export { PartnersTable } from './partnersTable/partnersTable';
-export { SignInModal } from './signInModal/signInModal';
+export { default as AboutIndex } from './AboutIndex';
+export { default as Footer } from './Footer';
+export { default as Layout } from './Layout';
+export { default as Map } from './Map';
+export { default as Modal } from './Modal';
+export { default as Nav } from './Nav';
+export { default as PartnersTable } from './PartnersTable';
+export { default as SignInModal } from './SignInModal';
+export { default as SignupForm } from './SignupForm';
+export { default as Table } from './Table';
+export { default as UsersLookup } from './UsersLookup';
