--- components/map/index.tsx
@@ -1,1 +1,0 @@
-export { default as Map } from './map';
