--- components/modal/index.tsx
@@ -1,1 +1,0 @@
-export { Modal } from './modal';
