--- components/nav/index.tsx
@@ -1,1 +1,0 @@
-export { Nav } from './Nav';
