--- components/partnersTable/index.tsx
@@ -1,1 +1,0 @@
-export { PartnersTable } from './partnersTable';
