--- components/signInModal/index.tsx
@@ -1,1 +1,0 @@
-export { SignInModal } from './signInModal';
