--- components/signupForm/index.tsx
@@ -1,1 +1,0 @@
-export { SignupForm } from './signupForm';
