--- components/usersLookup/index.tsx
@@ -1,1 +1,0 @@
-export { UsersLookup } from './usersLookup';
